fixed tests

+++ Receipt/Receipt.cpp
@@ -16,7 +16,7 @@ void Receipt::print_receipt()
     std::string no_tax_ending = "---- without tax: " + std::to_string(price_without_tax.get_whole_part())+"."+std::to_string(price_without_tax.get_cents())+price_without_tax.currency_name+"\n";
 
     receipt_interface.print(header);
-    for(int i=0;i<names.size();i++)
+    for(unsigned long int i=0;i<names.size();i++)
     {
         receipt_interface.print("---- "+names[i]+": "+std::to_string(prices[i].get_whole_part())+":"+std::to_string(prices[i].get_cents())+prices[i].currency_name+"\n");
     }
